Consolidate away KernelPluginExtensions
We currently have multiple Extension types living in the Microsoft.SemanticKernel namespace, including `KernelPluginExtensions` that's in the core assembly and `IKernelPluginExtensions` that's in the abstractions assembly.  This:
- Consolidates the two remaining extensions classes in Core into a single one by moving the two members from `KernelPluginExtensions` to `KernelExtensions`, helping to unclutter the namespace and reduce similarly-named extension types.
- Then renames `IKernelPluginExtensions` to `KernelPluginExtensions`, to avoid the type looking like an interface, now that the good name is available.
- Moves `KernelExtensions.InvokeAsync<T>` down to being instance members of Kernel.
